wip(segacd): subchannel Q synth + CDC DMA path + Mednafen reference investigation
Checkpoint save — subchannel Q synthesis, CDC PRG-RAM DMA path wiring,
segacd_bus $FF8100 uint8 truncation fix, and the pivot to Mednafen as an
independent reference emulator (PicoDrive can't reach Sonic CD gameplay
either). Not yet playable — sub program load / boot handshake still open.

+++ Core/Src/porting/segacd/segacd_bus.c
@@ -102,7 +102,7 @@ static unsigned int sub_ff_read8(unsigned int address)
      * status register within POLL_CYCLES(52) sub cycles. At POLL_LIMIT(16)
      * repeats, mark the sub as idle — segacd_run_sub skips its timeslices
      * until a write (segacd_poll_wake) or pending IRQ re-arms it. */
-    uint8_t reg = (uint8_t)(off & (SEGACD_GA_REGS - 1));
+    unsigned int reg = off & (SEGACD_GA_REGS - 1);
     SGA_RD(reg);
     if (!SCD.sub_idle) {
         if (reg == SCD.poll_reg) {
